Minor fixes on categories
